+++ sentry/src/main/java/io/sentry/exception/InvalidSentryTraceHeaderException.java
@@ -2,6 +2,7 @@ package io.sentry.exception;
 
 import io.sentry.SentryTraceHeader;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Thrown when {@link SentryTraceHeader} fails to create because of the invalid value of the
@@ -12,7 +13,12 @@ public final class InvalidSentryTraceHeaderException extends Exception {
   private final @NotNull String sentryTraceHeader;
 
   public InvalidSentryTraceHeaderException(final @NotNull String sentryTraceHeader) {
-    super("sentry-trace header does not conform to expected format: " + sentryTraceHeader);
+    this(sentryTraceHeader, null);
+  }
+
+  public InvalidSentryTraceHeaderException(
+      final @NotNull String sentryTraceHeader, final @Nullable Throwable cause) {
+    super("sentry-trace header does not conform to expected format: " + sentryTraceHeader, cause);
     this.sentryTraceHeader = sentryTraceHeader;
   }