+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Make SentryAppender non-final for Log4j2 and Logback (#1603) 
+* Fix: Do not throw IAE when tracing header contain invalid trace id (#1605)
 
 ## 5.1.0-beta.4